+++ src/items/mask.jl
@@ -1,17 +1,27 @@
 """
-    MaskMulti(a, [classes])
+    MaskMulti(a, [classes]; interpolate=BSpline(Constant()), extrapolate=Flat())
 
-An `N`-dimensional multilabel mask with labels `classes`.
+An `N`-dimensional multilabel mask with labels `classes`. Optionally, the
+interpolation and extrapolation method can be provided. Interpolation here
+refers to how the values of projected pixels that fall into the transformed
+content bounds are calculated. Extrapolation refers to how to assign values
+that fall outside the projected content bounds. The default is nearest neighbor
+interpolation and flat extrapolation of the edges into new regions.
+
+!!! info
+    The `Interpolations` package provides numerous methods for use with
+    the `interpolate` and `extrapolate` keyword arguments.  For instance,
+    `BSpline(Linear())` and `BSpline(Constant())` provide linear and nearest
+    neighbor interpolation, respectively. In addition `Flat()`, `Reflect()` and
+    `Periodic()` boundary conditions are available for extrapolation.
 
 ## Examples
 
-{cell=MaskMulti}
 ```julia
 using DataAugmentation
 
 mask = MaskMulti(rand(1:3, 100, 100))
 ```
-{cell=MaskMulti}
 ```julia
 showitems(mask)
 ```
@@ -20,19 +30,30 @@ struct MaskMulti{N, T<:Integer, U} <: AbstractArrayItem{N, T}
     data::AbstractArray{T, N}
     classes::AbstractVector{U}
     bounds::Bounds{N}
+    interpolate::Interpolations.InterpolationType
+    extrapolate::ImageTransformations.FillType
 end
 
+function MaskMulti(
+    data::AbstractArray{T,N},
+    classes::AbstractVector{U},
+    bounds::Bounds{N};
+    interpolate::Interpolations.InterpolationType=BSpline(Constant()),
+    extrapolate::ImageTransformations.FillType=Flat(),
+) where {N, T<:Integer, U}
+    return MaskMulti(data, classes, bounds, interpolate, extrapolate)
+end
 
-function MaskMulti(a::AbstractArray, classes = unique(a))
+function MaskMulti(a::AbstractArray, classes = unique(a); kwargs...)
     bounds = Bounds(size(a))
     minimum(a) >= 1 || error("Class values must start at 1")
-    return MaskMulti(a, classes, bounds)
+    return MaskMulti(a, classes, bounds; kwargs...)
 end
 
-MaskMulti(a::AbstractArray{<:Gray{T}}, args...) where T = MaskMulti(reinterpret(T, a), args...)
-MaskMulti(a::AbstractArray{<:Normed{T}}, args...) where T = MaskMulti(reinterpret(T, a), args...)
-MaskMulti(a::IndirectArray, classes = a.values, bounds = Bounds(size(a))) =
-    MaskMulti(a.index, classes, bounds)
+MaskMulti(a::AbstractArray{<:Gray{T}}, args...; kwargs...) where T = MaskMulti(reinterpret(T, a), args...; kwargs...)
+MaskMulti(a::AbstractArray{<:Normed{T}}, args...; kwargs...) where T = MaskMulti(reinterpret(T, a), args...; kwargs...)
+MaskMulti(a::IndirectArray, classes = a.values, bounds = Bounds(size(a)); kwargs...) =
+    MaskMulti(a.index, classes, bounds; kwargs...)
 
 Base.show(io::IO, mask::MaskMulti{N, T}) where {N, T} =
     print(io, "MaskMulti{$N, $T}() with size $(size(itemdata(mask))) and $(length(mask.classes)) classes")
@@ -41,12 +62,15 @@ Base.show(io::IO, mask::MaskMulti{N, T}) where {N, T} =
 getbounds(mask::MaskMulti) = mask.bounds
 
 
-function project(P, mask::MaskMulti, bounds::Bounds)
-    a = itemdata(mask)
-    etp = mask_extrapolation(a)
-    res = warp(etp, inv(P), bounds.rs)
+function project(P, mask::MaskMulti{N, T, U}, bounds::Bounds{N}) where {N, T, U}
+    res = warp(
+        itemdata(mask),
+        inv(P),
+        bounds.rs;
+        method=mask.interpolate,
+        fillvalue=mask.extrapolate)
     return MaskMulti(
-        res,
+        convert.(T, res),
         mask.classes,
         bounds
     )
@@ -57,7 +81,7 @@ function project!(bufmask::MaskMulti, P, mask::MaskMulti, bounds)
     a = OffsetArray(parent(itemdata(bufmask)), bounds.rs)
     warp!(
         a,
-        mask_extrapolation(itemdata(mask)),
+        box_extrapolation(itemdata(mask); method=mask.interpolate, fillvalue=mask.extrapolate),
         inv(P),
     )
     return MaskMulti(
@@ -80,19 +104,29 @@ end
 # ## Binary masks
 
 """
-    MaskBinary(a)
+    MaskBinary(a; interpolate=BSpline(Constant()), extrapolate=Flat())
 
-An `N`-dimensional binary mask.
+An `N`-dimensional binary mask. Optionally, the interpolation and extrapolation
+method can be provided. Interpolation here refers to how the values of
+projected pixels that fall into the transformed content bounds are calculated.
+Extrapolation refers to how to assign values that fall outside the projected
+content bounds. The default is nearest neighbor interpolation and flat
+extrapolation of the edges into new regions.
+
+!!! info
+    The `Interpolations` package provides numerous methods for use with
+    the `interpolate` and `extrapolate` keyword arguments.  For instance,
+    `BSpline(Linear())` and `BSpline(Constant())` provide linear and nearest
+    neighbor interpolation, respectively. In addition `Flat()`, `Reflect()` and
+    `Periodic()` boundary conditions are available for extrapolation.
 
 ## Examples
 
-{cell=MaskMulti}
 ```julia
 using DataAugmentation
 
 mask = MaskBinary(rand(Bool, 100, 100))
 ```
-{cell=MaskMulti}
 ```julia
 showitems(mask)
 ```
@@ -100,10 +134,17 @@ showitems(mask)
 struct MaskBinary{N} <: AbstractArrayItem{N, Bool}
     data::AbstractArray{Bool, N}
     bounds::Bounds{N}
+    interpolate::Interpolations.InterpolationType
+    extrapolate::ImageTransformations.FillType
 end
 
-function MaskBinary(a::AbstractArray{Bool, N}, bounds = Bounds(size(a))) where N
-    return MaskBinary(a, bounds)
+function MaskBinary(
+    a::AbstractArray,
+    bounds = Bounds(size(a));
+    interpolate::Interpolations.InterpolationType=BSpline(Constant()),
+    extrapolate::ImageTransformations.FillType=Flat(),
+)
+    return MaskBinary(a, bounds, interpolate, extrapolate)
 end
 
 Base.show(io::IO, mask::MaskBinary{N}) where {N} =
@@ -112,19 +153,21 @@ Base.show(io::IO, mask::MaskBinary{N}) where {N} =
 getbounds(mask::MaskBinary) = mask.bounds
 
 function project(P, mask::MaskBinary, bounds::Bounds)
-    etp = mask_extrapolation(itemdata(mask))
-    return MaskBinary(
-        warp(etp, inv(P), bounds.rs),
-        bounds,
-    )
+    res = warp(
+        itemdata(mask),
+        inv(P),
+        bounds.rs;
+        method=mask.interpolate,
+        fillvalue=mask.extrapolate)
+    return MaskBinary(convert.(Bool, res), bounds)
 end
 
 
 function project!(bufmask::MaskBinary, P, mask::MaskBinary, bounds)
     a = OffsetArray(parent(itemdata(bufmask)), bounds.rs)
-    res = warp!(
+    warp!(
         a,
-        mask_extrapolation(itemdata(mask)),
+        box_extrapolation(itemdata(mask); method=mask.interpolate, fillvalue=mask.extrapolate),
         inv(P),
     )
     return MaskBinary(
@@ -135,16 +178,4 @@ end
 
 function showitem!(img, mask::MaskBinary)
     showimage!(img, colorview(Gray, itemdata(mask)))
-end
-# ## Helpers
-
-
-function mask_extrapolation(
-        mask::AbstractArray{T};
-        t = T,
-        degree = Constant(),
-        boundary = Flat()) where T
-    itp = interpolate(t, t, mask, BSpline(degree))
-    etp = extrapolate(itp, Flat())
-    return etp
 end