+++ test/imports.jl
@@ -5,6 +5,11 @@ using TestSetExtensions
 using CoordinateTransformations
 using Colors
 using FixedPointNumbers: N0f8
+using LinearAlgebra
+using Rotations
+using Statistics
+using OffsetArrays
+import Interpolations
 
 
 using DataAugmentation: Item, Transform, getrandstate, itemdata, setdata, ComposedProjectiveTransform,